some more fixes
git-svn-id: https://svn.forgerock.org/openam/branches/openam_10.2.0_xacml3_JAS@4774 0f4defcf-c51a-4c67-9f44-6fb5eba73c5d

+++ openam/openam-entitlements/src/main/java/org/forgerock/openam/xacml/v3/model/FunctionArgument.java
@@ -229,23 +229,23 @@ public abstract class FunctionArgument  {
         }
         return (String) fArg.getValue(pip);
     }
-    public Integer asDayTimeDuration(XACMLEvalContext pip) throws XACML3EntitlementException {
+    public Long asDayTimeDuration(XACMLEvalContext pip) throws XACML3EntitlementException {
 
         FunctionArgument fArg = getDataContainer(pip);
 
         if (!fArg.dataType.isType(DataType.Type.XACMLDAYTIMEDURATIONTYPE)) {
             throw new IndeterminateException("type conflict");
         }
-        return (Integer) fArg.getValue(pip);
+        return (Long) fArg.getValue(pip);
     }
-    public Integer asYearMonthDuration(XACMLEvalContext pip) throws XACML3EntitlementException {
+    public Long asYearMonthDuration(XACMLEvalContext pip) throws XACML3EntitlementException {
 
         FunctionArgument fArg = getDataContainer(pip);
 
         if (!fArg.dataType.isType(DataType.Type.XACMLYEARMONTHDURATIONTYPE)) {
             throw new IndeterminateException("type conflict");
         }
-        return (Integer) fArg.getValue(pip);
+        return (Long) fArg.getValue(pip);
     }
     public String asX500Name(XACMLEvalContext pip) throws XACML3EntitlementException {
